+++ src/app/oauth/callback/page.tsx
@@ -1,44 +1,13 @@
 'use client';
 
 import { SpinnerIcon } from '@/components/icons';
-import { AUTH_PATH } from '@/constants';
-import { useMe } from '@/features/auth/hooks';
-import { accessTokenAtom, signupTokenAtom } from '@/store';
-import { useSetAtom } from 'jotai';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
+import { LoginCallbackContent } from '@/features/auth/components';
+import { Suspense } from 'react';
 
 export default function LoginCallbackPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const message = searchParams.get('message');
-  const setSignupToken = useSetAtom(signupTokenAtom);
-  const setAccessToken = useSetAtom(accessTokenAtom);
-
-  const { mutate: getMe } = useMe();
-
-  useEffect(() => {
-    if (message === 'additionalInfoRequired') {
-      setSignupToken(searchParams.get('signupToken'));
-      router.push(AUTH_PATH.SIGN_UP);
-    } else if (message === 'loginSuccess') {
-      const accessToken = searchParams.get('accessToken') as string;
-
-      setAccessToken(accessToken);
-      getMe(accessToken);
-    } else {
-      // 로그인 실패 처리 Toast
-      router.push(AUTH_PATH.LOGIN);
-    }
-  }, [message]);
   return (
-    <section className="flex h-screen w-full items-center justify-center">
-      <SpinnerIcon
-        width={32}
-        height={32}
-        fill="var(--color-blue)"
-        className="animate-spin"
-      />
-    </section>
+    <Suspense fallback={<SpinnerIcon />}>
+      <LoginCallbackContent />
+    </Suspense>
   );
 }
