+++ src/features/auth/components/index.ts
@@ -1,2 +1,3 @@
 export * from './auth-header';
+export * from './login-callback-content';
 export * from './signup';
