+++ src/features/auth/components/login-callback-content/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginCallbackContent';
